Add files via upload
This is a react project for beginners. Implementation of different ipod functionality.

Features: Built a menu of options you see such as Settings, Games, Music etc
On clicking of the centre button it should go inside the selected item
Clicking on the ok should hide the menu and open a screen of that particular option, like settings, games etc
Click and hold on the circular menu and move the mouse in a circular fashion inside the menu boundary (white circle is the menu) to navigate in the iPod menu
Clicking on “Menu” should take you back to the main menu
create sub menu inside menu like in Music option, it has sub menu with options “All songs”, “Artists” etc.